Create SwiftPluginServerSupport
