+++ tools/swift-plugin-server/Sources/swift-wasm-plugin-server/JSCWasmPlugin.swift
@@ -11,6 +11,7 @@
 //===----------------------------------------------------------------------===//
 
 import JavaScriptCore
+import SwiftPluginServerSupport
 
 // returns: Promise<(ArrayBuffer) => ArrayBuffer>
 private let js = """
